+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add functions
+    - `Serialize.toJson`
+    - `Serialize.toJsonPretty`
 
-## 0.1.0 - 
+## 0.1.0 -
 - Initial implementation